requirements.txt
pyrogram
tgcrypto
python-dotenv
psutil
requests